backend/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
django-tables
lorem